feat: add feature to guess fruit
add the feature to display a list of fruit on the screen for the user to predict
then we take the user's input and determine whether they are correct or not. The
system asks the user's name to welcome them to the system then confirms whether
the user wants to play fruit or country.

Closes: #4